save places

+++ ansible/ceph/teardown.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 DISK="/dev/${1}"
+if [ -z ${DISK} ] ; then
+    echo "Usage: $0 <disk>"
+    echo "where <disk> is like hda or sdb"
+    exit 1
+fi
 
 # Zap the disk to a fresh, usable state (zap-all is important, b/c MBR has to be clean)
 
@@ -15,12 +20,13 @@ blkdiscard $DISK
 CEPH_ID="$(lsblk ${DISK} | grep ceph | awk '{print $1}'  | sed -e 's/└─//g')"
 # These steps only have to be run once on each node
 # If rook sets up osds using ceph-volume, teardown leaves some devices mapped that lock the disks.
-find /dev/mapper -type b -name "${CEPH_ID}" -exec dmsetup remove {} +
-find /dev/ -type b -name "${CEPH_ID}" -delete
-
-# ceph-volume setup can leave ceph-<UUID> directories in /dev and /dev/mapper (unnecessary clutter)
-rm -rf /dev/${CEPH_ID}
+if [ -z ${CEPH_ID} ] ; then
+    eho "could not get lsblk disks labeled as CEPH"
+    exit 255
+fi
+dmsetup remove /dev/mapper/${CEPH_ID}
 rm -rf /dev/mapper/${CEPH_ID}
+rm -rf  /dev/${CEPH_ID}
 
 # Inform the OS of partition table changes
 partprobe $DISK
